FixedMinimapPlusLeaks

+++ src/get_next_line_utils.c
@@ -6,7 +6,7 @@
 /*   By: pdrettas <pdrettas@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/24 18:40:06 by pdrettas          #+#    #+#             */
-/*   Updated: 2024/11/24 18:40:09 by pdrettas         ###   ########.fr       */
+/*   Updated: 2025/07/14 14:27:00 by itsiros          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 	return (src_length);
 }
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
